Add Private Key Secret
Aligned with what has already been implemented in the
result service, a secret is added for providing a
private key to the message broker service.

NOTE: This is known to be code duplication. However,
since there's a major overhaul planned where variables
can get re-used within multiple sub-charts, this can
be considered "good enough" for now.